Made Changes to part1.cpp
Created the test file finalp1.txt
Created outout file output.txt
Made some changeds to the while loop instead of reading by line it is reading by char.
there is a bug when it read the last word. The text that is in the output is currently what the code outputs